Fix /playtime date formats

+++ plugins/sensor/src/main/java/com/github/maizenalegal/sensor/identity/Username.kt
@@ -13,7 +13,7 @@ object Username {
         val createdAt: String,
         val firstJoin: String?,
         val lastSeen: String?,
-        val playedSeconds: Int,
+        val playedSeconds: Long,
         val joinCount: Int,
         val isOnline: Boolean
     )
